requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
yt-dlp
opencv-python
numpy
Pillow
